ONNX and SPDLOG
Move to ONNX as a DNN runtime. ONNX is a framework-agnostic standard,
and the runtime is stable / easy to deploy. This means: using pre-
compiled Tensorflow was a big liability and is *not* futureproof (and is
the reason that SVD was exclusive for Linux for some years as friendly
folks from RWTH Aachen robotics provided such precompiled binaries, and
my attempts to compile newer TF versions on windows failed miserably)

What changed in the code:
- most of tensorflow logic was encapsulated with TensorHelper classes,
actual changes where not dramatic. The biggest change is memory
management, which moved from TF on ONNX (SVD).
- loading / parsing of ONNX files

SPDLOG: updated from very old (2017) version to 1.15 (2026), which
required some refactoring.

Changes for using SVD:
- SVD now requiers .onnx files instead of binary Tensorflow files.
- therefore, onnx needs to be exported (in Python) after training

Current state:
SVD compiles without tensorflow (and without GPU as backend for ONNX)
still some things to do!

+++ SVDModel/Predictor/dnnshell.cpp
@@ -140,11 +140,11 @@ void DNNShell::doWork(Batch *batch)
 
 
     if (batch->state()!=Batch::Fill)
-        lg->error("Batch {} [{}] is in the wrong state {}, size: {}", batch->packageId(), static_cast<void*>(batch), batch->state(), batch->usedSlots());
+        lg->error("Batch {} [{}] is in the wrong state {}, size: {}", batch->packageId(), static_cast<void*>(batch), static_cast<int>(batch->state()), batch->usedSlots());
 
     batch->changeState(Batch::DNNInference);
     mProcessing++;
-    lg->debug("DNNShell: received package {}. Starting DNN (batch: {}, state: {}, active threads now: {}, #processing: {}) ", batch->packageId(), static_cast<void*>(batch), batch->state(), mThreads->activeThreadCount(), mProcessing);
+    lg->debug("DNNShell: received package {}. Starting DNN (batch: {}, state: {}, active threads now: {}, #processing: {}) ", batch->packageId(), static_cast<void*>(batch), static_cast<int>(batch->state()), mThreads->activeThreadCount(), static_cast<int>(mProcessing));
 
     QtConcurrent::run( mThreads,
                        [](DNNShell *shell, Batch *batch, DNN *dnn){
